+++ CHANGELOG.md
@@ -1,14 +1,22 @@
 # Changelog
 
+## 2023-02-10
+
 ### Changed
 
-- [CL-2673] In-platform survey UI improvements
+- Tooltips for image uploaders in project and folder forms.
+- [CL-737] Show login screen after successful password reset
 
-## Next release
+## 2023-02-09
 
 ### Added
 
 - Report builder templates now automatically detect ideation and native survey projects and add different widgets based on that
+- [CL-2739] Added basic posthog integration
+
+### Changed
+
+- [CL-2673] In-platform survey UI improvements
 
 ## 2023-02-03